Finalize new landing page.
