+++ client/components/Posts/BookPost/BookPost.js
@@ -1,5 +1,6 @@
 /**
- * @file React component for a textbook posting on the webapp.
+ * @file BookPost.js
+ * @description React component for a textbook posting on the webapp.
  * @author Duncan Grubbs <duncan.grubbs@gmail.com>
  * @version 0.0.4
  */
@@ -7,36 +8,40 @@
 import React, { Component } from 'react';
 import propTypes from 'prop-types';
 
-import FetchService from '../../services/FetchService';
-import AuthService from '../../services/AuthService';
+import FetchService from '../../../services/FetchService';
+import AuthService from '../../../services/AuthService';
 
-class CartBookPost extends Component {
+import './BookPost.css';
+
+class BookPost extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      id: '',
+      inCart: false,
     };
 
-    this.removeFromCart = this.removeFromCart.bind(this);
+    this.addToCart = this.addToCart.bind(this);
   }
 
   componentDidMount() {
-    this._setID();
+    if (this.props.status === 42) {
+      this._setInCart();
+    }
   }
 
-  _setID() {
-    this.setState({ id: this.props.id });
+  _setInCart() {
+    this.setState({ inCart: true });
   }
 
-  removeFromCart() {
+  addToCart() {
     const AUTH = new AuthService();
-    FetchService.POST('/api/user/removeFromCart', {
-      bookID: this.state.id,
+    FetchService.POST('/api/user/addToCart', {
       token: AUTH.getToken(),
+      bookID: this.props.id,
     })
       .then(() => {
-        window.location.reload();
+        this.setState({ inCart: true });
       });
   }
 
@@ -50,18 +55,28 @@ class CartBookPost extends Component {
         </div>
         <div className="vertical-line" />
         <div className="leftBP">
-          <div>
-            <span className="condition">{this.props.condition}</span>
-             for <span className="price">${this.props.price}</span>
-          </div>
           <span className="comments"><i>{this.props.comments || 'No comments'}</i></span>
         </div>
         <div className="rightBP">
-          <button
-            className="button"
-            onClick={this.removeFromCart}
-          >Remove
-          </button>
+          <div>
+            <span className="condition">{this.props.condition}</span>
+             - <span className="price">${this.props.price}</span>
+          </div>
+          {
+            !this.state.inCart &&
+            <button
+              className="button"
+              onClick={this.addToCart}
+            >Add to Cart
+            </button>
+          }
+          {
+            this.state.inCart &&
+            <button
+              className="button green"
+            >Added to Cart
+            </button>
+          }
         </div>
       </div>
     );
@@ -69,14 +84,15 @@ class CartBookPost extends Component {
 }
 
 // Props validation
-CartBookPost.propTypes = {
+BookPost.propTypes = {
   comments: propTypes.string,
   condition: propTypes.string.isRequired,
   edition: propTypes.number.isRequired,
   id: propTypes.string.isRequired,
   name: propTypes.string.isRequired,
   price: propTypes.number.isRequired,
+  status: propTypes.number,
   subject: propTypes.string.isRequired,
 };
 
-export default CartBookPost;
+export default BookPost;
