+++ client/routes/Home/Home.js
@@ -10,7 +10,7 @@ import './Home.css';
 import TopBar from '../../components/TopBar/TopBar';
 import SideNav from '../../components/SideNav/SideNav';
 import Notification from '../../components/Notification/Notification';
-import BookPost from '../../components/BookPost/BookPost';
+import BookPost from '../../components/Posts/BookPost/BookPost';
 
 import FetchService from '../../services/FetchService';
 import AuthService from '../../services/AuthService';
